backend/requirements.txt
fastapi==0.95.0
openai==0.23.0
passlib==1.7.4
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
uvicorn==0.21.1
python-multipart==0.0.6
